added chart.js to utils folder, results.html template, and /results route to home routes.
