--- chart.js
@@ -1,41 +1,0 @@
-import { Chart, registerables } from "chart.js";
-Chart.register(...registerables);
-
-var click = document.getElementById("chartHolder");
-const ctx = document.getElementById("resultsChart");
-const Chart = require("chart.js");
-
-
-  const myChart = new Chart(ctx, {
-    type: "doughnut",
-    labels: ["Proposal 1", "Proposal 2", "Proposal3", "Proposal4", "Proposal5"],
-    data: {
-      labels: labels,
-      datasets: [
-        {
-          label: "Number of Votes",
-          data: [12, 19, 3, 5, 2, 3],
-          backgroundColor: [
-            "rgba(255, 99, 132, 0.2)",
-            "rgba(54, 162, 235, 0.2)",
-            "rgba(255, 206, 86, 0.2)",
-            "rgba(75, 192, 192, 0.2)",
-            "rgba(153, 102, 255, 0.2)",
-          ],
-          borderColor: ["rgba(255, 99, 132, 1)"],
-          borderWidth: 1,
-        },
-      ],
-    },
-    options: {},
-  });
-}
-
-
-// const config = {
-//   type: "donut",
-//   data: data,
-//   options: {}
-// }
-
-module.exports = myChart
