+++ views/layouts/chart.html
