+++ views/layouts/timer.html
@@ -6,6 +6,8 @@
   <button onclick="countdownTimeStart()">Start Timer</button>
 </div> -->
 
+<!-- -------------------------------- -->
+
 <!-- Full code demo - open file in browser to see in operation:  -->
 <!-- Script file is added to utils/helpers.js -->
